# talent-match-system/docs/api/api-contract.yaml
openapi: 3.0.0
info:
  title: 人才职位智能匹配与能力图谱系统 API
  version: 1.0.0
  description: 基于AI的人才职位智能匹配与能力图谱系统API接口文档
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/user/login:
    post:
      summary: 用户登录
      description: 手机号密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: 手机号
                password:
                  type: string
                  description: 密码
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT访问令牌
                  token_type:
                    type: string
                    description: 令牌类型
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      role:
                        type: string
                      roles:
                        type: array
                        items:
                          type: string
        '401':
          description: 认证失败
...
  /api/admin/users:
    get:
      summary: 获取用户列表（管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        phone:
                          type: string
                        roles:
                          type: array
                          items:
                            type: string
                        created_at:
                          type: string
                          format: date-time
  /api/admin/users/{user_id}:
    get:
      summary: 获取用户详情（管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
        '404':
          description: 用户不存在

    delete:
      summary: 删除用户（管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
        '404':
          description: 用户不存在
  /api/admin/users/{user_id}/roles:
    put:
      summary: 更新用户角色（管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 更新成功
  /api/admin/users/{user_id}/promote:
    post:
      summary: 提升为管理员
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 提升成功
  /api/admin/stats:
    get:
      summary: 获取系统统计（管理员）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  total_resumes:
                    type: integer
                  total_jobs:
                    type: integer
                  total_matches:
                    type: integer
                  job_seeker_count:
                    type: integer
                  company_count:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT